requirements.txt
anyio==3.6.2
autopep8==2.0.2
backports.zoneinfo==0.2.1
click==8.1.3
fastapi==0.95.0
h11==0.14.0
httptools==0.5.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg==3.1.8
psycopg-binary==3.1.8
psycopg-pool==3.1.7
pycodestyle==2.10.0
pydantic==1.10.6
python-dotenv==1.0.0
PyYAML==6.0
sniffio==1.3.0
starlette==0.26.1
tomli==2.0.1
typing_extensions==4.5.0
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
